validate functions for genesis subtypes
